% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMGE.AMMI.R
\name{AMGE.AMMI}
\alias{AMGE.AMMI}
\title{Sum Across Environments of GEI Modelled by AMMI}
\usage{
AMGE.AMMI(model, n, alpha = 0.05)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}.}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}
}
\value{

}
\description{
\code{AMGE.AMMI} computes the Sum Across Environments of GEI Modelled by AMMI
(AMGE) (Sneller et al., 1997) considering all significant interaction
principal components (IPCs) in the AMMI model. Using AMGE, the Yield stability Index
(YSI) is also calculated.
}
\details{
The Sum Across Environments of GEI Modelled by AMMI (\eqn{AMGE}) is computed
as follows:

\deqn{AMGE = \sum_{j=1}^{E} \sum_{n=1}^{N'} \lambda_{n} \gamma_{in}
\delta_{jn}}

Where, \eqn{N'} is the number of significant IPCs (number of IPC that were
retained in the AMMI model via F tests); \eqn{\lambda_{n}} is the is the
singular value for IPC \eqn{n} and correspondingly \eqn{\lambda_{n}^{2}}  is
its eigen value; \eqn{\gamma_{in}} is the eigenvector value for \eqn{i}th
genotype; and \eqn{\delta_{jn}}  is the eigenvector value for \eqn{j}th
environment.

The Yield Stability Index (\eqn{YSI}) is computed as follows:

\deqn{YSI = R_{AMGE} + R_{Y}}

Where, \eqn{R_{AMGE}} is the SIPC rank of the genotype and \eqn{R_{Y}} is the
mean yield rank of the genotype.
}
\references{
\insertRef{sneller_repeatability_1997}{AMMIStbP}
}
