% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammistability.R
\name{ammistability}
\alias{ammistability}
\title{Estimate multiple AMMI model Stability Parameters}
\usage{
ammistability(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"),
  a = 1, AMGE = TRUE, ASI = TRUE, ASV = TRUE, ASTAB = TRUE,
  AVAMGE = TRUE, DA = TRUE, DZ = TRUE, EV = TRUE, FA = TRUE,
  MASI = TRUE, MASV = TRUE, SIPC = TRUE, ZA = TRUE,
  force.grouping = TRUE, line.size = 1, line.alpha = 0.5,
  line.col = NULL, point.size = 1, point.alpha = 0.5, point.col = NULL,
  text.size = 2)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}).}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{AMGE}{If \code{TRUE}, computes AMGE (see \strong{Details}). Default is
\code{TRUE}.}

\item{ASI}{If \code{TRUE}, computes ASI (see \strong{Details}). \code{n = 2}
will beused in this case. Default is \code{TRUE}.}

\item{ASV}{If \code{TRUE}, computes ASV (see \strong{Details}). \code{n = 2}
will beused in this case. Default is \code{TRUE}.}

\item{ASTAB}{If \code{TRUE}, computes ASTAB (see \strong{Details}). Default
is \code{TRUE}.}

\item{AVAMGE}{If \code{TRUE}, computes AVAMGE (see \strong{Details}). Default
is \code{TRUE}.}

\item{DA}{If \code{TRUE}, computes DA (see \strong{Details}). Default is
\code{TRUE}.}

\item{DZ}{If \code{TRUE}, computes DZ (see \strong{Details}). Default is
\code{TRUE}.}

\item{EV}{If \code{TRUE}, computes EV (see \strong{Details}). Default is
\code{TRUE}.}

\item{FA}{If \code{TRUE}, computes FA (see \strong{Details}). Default is
\code{TRUE}.}

\item{MASI}{If \code{TRUE}, computes MASI (see \strong{Details}). Default is
\code{TRUE}.}

\item{MASV}{If \code{TRUE}, computes MASV (see \strong{Details}). Default is
\code{TRUE}.}

\item{SIPC}{If \code{TRUE}, computes SIPC (see \strong{Details}). Default is
\code{TRUE}.}

\item{ZA}{If \code{TRUE}, computes ZA (see \strong{Details}). Default is
\code{TRUE}.}

\item{force.grouping}{If \code{TRUE}, genotypes will be considered as a
grouping variable for plotting the slopegraphs. (Each genotype will be
represented by a different colour in the slopegraphs). Default is
\code{TRUE}.}

\item{line.size}{Size of lines plotted in the slopegraphs. Must be numeric.}

\item{line.alpha}{Transparency of lines plotted in the slopegraphs. Must be
numeric.}

\item{line.col}{Default is \code{TRUE}. Overrides colouring by
\code{force.grouping} argument.}

\item{point.size}{Size of points plotted in the slopegraphs. Must be numeric.}

\item{point.alpha}{Transparency of points plotted in the slopegraphs. Must be
numeric.}

\item{point.col}{Default is \code{TRUE}. Overrides colouring by
\code{force.grouping} argument.}

\item{text.size}{Size of text annotations plotted in the slopegraphs. Must be
numeric.}
}
\description{
\code{ammistability} computes multiple stability parameters from an AMMI
model. Further, the corresponding Simultaneous Selection Indices for Yield
and Stability (SSI) are also calculated according to the argument
\code{ssi.method}. From the results, correlation between the computed indices
will also be computed. The resulting correlation matrices will be plotted as
correlograms. For visual comparisons of ranks of genotypes for different
indices, slopegraphs will also be generated by this function.
}
\details{
\code{ammistability} computes the following stability parameters from an AMMI
model.

\describe{ \item{\link[AMMIStbP:AMGE.AMMI]{Sum Across Environments of GEI
Modelled by AMMI (AMGE)}}{Sneller et al., 1997}
\item{\link[AMMIStbP:ASI.AMMI]{AMMI Stability Index (ASI)}}{Jambhulkar et
al., 2014; Jambhulkar et al., 2015; Jambhulkar et al., 2017}
\item{\link[AMMIStbP:MASV.AMMI]{AMMI Stability Value (ASV)}}{Purchase 1997;
Purchase et al., 1999; Purchase et al., 2000}
\item{\link[AMMIStbP:ASTAB.AMMI]{AMMI Based Stability Parameter (ASTAB)}}{Rao
and Prabhakaran, 2005} \item{\link[AMMIStbP:AVAMGE.AMMI]{Sum Across
Environments of Absolute Value of GEI Modelled by AMMI (AVAMGE)}}{Zali et
al., 2012} \item{\link[AMMIStbP:DA.AMMI]{Annicchiarico's D Parameter
(DA)}}{Annicchiarico, 1997} \item{\link[AMMIStbP:DZ.AMMI]{Zhang's D Parameter
(DZ)}}{Zhang, 1998} \item{\link[AMMIStbP:EV.AMMI]{Averages of the Squared
Eigenvector Values (EV)}}{Zobel, 1994}
\item{\link[AMMIStbP:FA.AMMI]{Stability Measure Based on Fitted AMMI Model
(FA)}}{Raju, 2002} \item{\link[AMMIStbP:MASI.AMMI]{Modified AMMI Stability
Index (MASI)}}{Ajay et al., 2018} \item{\link[AMMIStbP:MASV.AMMI]{Modified
AMMI Stability Value (MASV)}}{Zali et al., 2012}
\item{\link[AMMIStbP:SIPC.AMMI]{Sums of the Absolute Value of the IPC Scores
(SIPC)}}{Sneller et al., 1997} \item{\link[AMMIStbP:ZA.AMMI]{Absolute Value
of the Relative Contribution of IPCs to the Interaction (Za)}}{Zali et al.,
2012} }
}
\examples{
library(agricolae)
data(plrv)

# AMMI model
model <- with(plrv, AMMI(Locality, Genotype, Rep, Yield, console = FALSE))

ammistability(model)
}
\references{
\insertRef{zobel_stress_1994}{AMMIStbP}

\insertRef{annicchiarico_joint_1997}{AMMIStbP}

\insertRef{purchase_parametric_1997}{AMMIStbP}

\insertRef{sneller_repeatability_1997}{AMMIStbP}

\insertRef{zhang_analysis_1998}{AMMIStbP}

\insertRef{purchase_use_1999}{AMMIStbP}

\insertRef{purchase_genotype_2000}{AMMIStbP}

\insertRef{raju_study_2002}{AMMIStbP}

\insertRef{rao_use_2005}{AMMIStbP}

\insertRef{zali_evaluation_2012}{AMMIStbP}

\insertRef{jambhulkar_ammi_2014}{AMMIStbP}

\insertRef{jambhulkar_genotype_2015}{AMMIStbP}

\insertRef{jambhulkar_stability_2017}{AMMIStbP}
}
\seealso{
\code{\link[agricolae]{AMMI}}, \code{\link[AMMIStbP]{AMGE.AMMI}},
  \code{\link[AMMIStbP]{ASI.AMMI}}, \code{\link[AMMIStbP]{ASTAB.AMMI}},
  \code{\link[AVAMGE]{AMGE.AMMI}}, \code{\link[AMMIStbP]{DA.AMMI}},
  \code{\link[AMMIStbP]{DZ.AMMI}}, \code{\link[AMMIStbP]{EV.AMMI}},
  \code{\link[AMMIStbP]{FA.AMMI}}, \code{\link[AMMIStbP]{MASV.AMMI}},
  \code{\link[AMMIStbP]{SIPC.AMMI}}, \code{\link[AMMIStbP]{ZA.AMMI}},
  \code{\link[SSI]{SSI.AMMI}}
}
