% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASI.AMMI.R
\name{MASI.AMMI}
\alias{MASI.AMMI}
\title{Modified AMMI Stability Index}
\usage{
MASI.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"), a = 1)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}).}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[ammistability]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[ammistability]{SSI}}).}
}
\value{
A data frame with the following columns:  \item{MASI}{The MASI
  values.} \item{SSI}{The computed values of simultaneous selection index for
  yield and stability.} \item{rMASI}{The ranks of MASI values.} \item{rY}{The
  ranks of the mean yield of genotypes.} \item{means}{The mean yield of the
  genotypes.}

  The names of the genotypes are indicated as the row names of the data
  frame.
}
\description{
\code{MASI.AMMI} computes the Modified AMMI Stability Index (MASI)
\insertCite{ajay_modified_2018}{ammistability} from a modified formula of
AMMI Stability Index (ASI)
\insertCite{jambhulkar_ammi_2014,jambhulkar_genotype_2015,jambhulkar_stability_2017}{ammistability}.
Unlike ASI, MASI calculates stability value considering all significant
interaction principal components (IPCs) in the AMMI model. Using MASI, the
Simultaneous Selection Index for Yield and Stability (SSI) is also calculated
according to the argument \code{ssi.method}.
}
\details{
The Modified AMMI Stability Index (\mjseqn{MASI})
\insertCite{ajay_modified_2018}{ammistability} is computed as follows:

\mjsdeqn{MASI = \sqrt{ \sum_{n=1}^{N'} PC_{n}^{2} \times \theta_{n}^{2}}}

Where, \mjseqn{PC_{n}} are the scores of \mjseqn{n}th IPC; and
\mjseqn{\theta_{n}} is the percentage sum of squares explained by the
\mjseqn{n}th principal component interaction effect.
}
\examples{
library(agricolae)
data(plrv)

# AMMI model
model <- with(plrv, AMMI(Locality, Genotype, Rep, Yield, console = FALSE))

# ANOVA
model$ANOVA

# IPC F test
model$analysis

# Mean yield and IPC scores
model$biplot

# G*E matrix (deviations from mean)
array(model$genXenv, dim(model$genXenv), dimnames(model$genXenv))

# With default n (N') and default ssi.method (farshadfar)
MASI.AMMI(model)

# With n = 4 and default ssi.method (farshadfar)
MASI.AMMI(model, n = 4)

# With default n (N') and ssi.method = "rao"
MASI.AMMI(model, ssi.method = "rao")

# Changing the ratio of weights for Rao's SSI
MASI.AMMI(model, ssi.method = "rao", a = 0.43)

# ASI.AMMI same as MASI.AMMI with n = 2

a <- ASI.AMMI(model)
b <- MASI.AMMI(model, n = 2)

identical(a$ASI, b$MASI)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[agricolae]{AMMI}},
  \code{\link[ammistability]{ASI.AMMI}}, \code{\link[ammistability]{SSI}}
}
