% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{rankslopegraph}
\alias{rankslopegraph}
\title{Title}
\usage{
rankslopegraph(df, names, group, force.grouping = TRUE, line.size = 1,
  line.alpha = 0.5, line.col = NULL, point.size = 1, point.alpha = 0.5,
  point.col = NULL, text.size = 2, legend.position = "bottom")
}
\arguments{
\item{legend.position}{}
}
\description{
Title
}
