% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMGE.AMMI.R
\name{AMGE.AMMI}
\alias{AMGE.AMMI}
\title{Sum Across Environments of GEI Modelled by AMMI}
\usage{
AMGE.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"),
  a = 1)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}.}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[AMMIStbP]{SSI}}).}
}
\value{

}
\description{
\code{AMGE.AMMI} computes the Sum Across Environments of GEI Modelled by AMMI
(AMGE) (Sneller et al., 1997) considering all significant interaction
principal components (IPCs) in the AMMI model. Using AMGE, the Simultaneous
Selection Index for Yield and Stability (SSI) is also calculated according to
the argument \code{ssi.method}.
}
\details{
The Sum Across Environments of GEI Modelled by AMMI (\eqn{AMGE}) is computed
as follows:

\deqn{AMGE = \sum_{j=1}^{E} \sum_{n=1}^{N'} \lambda_{n} \gamma_{in}
\delta_{jn}}

Where, \eqn{N'} is the number of significant IPCs (number of IPC that were
retained in the AMMI model via F tests); \eqn{\lambda_{n}} is the is the
singular value for IPC \eqn{n} and correspondingly \eqn{\lambda_{n}^{2}}  is
its eigen value; \eqn{\gamma_{in}} is the eigenvector value for \eqn{i}th
genotype; and \eqn{\delta_{jn}}  is the eigenvector value for \eqn{j}th
environment.
}
\references{
\insertRef{sneller_repeatability_1997}{AMMIStbP}
}
\seealso{
\code{\link[AMMIStbP]{SSI}}
}
