% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMGE.AMMI.R
\name{AMGE.AMMI}
\alias{AMGE.AMMI}
\title{Sum Across Environments of GEI Modelled by AMMI}
\usage{
AMGE.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"),
  a = 1)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}).}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[AMMIStbP]{SSI}}).}
}
\value{
A data frame with the following columns:  \item{AMGE}{The AMGE
  values.} \item{SSI}{The computed values of simultaneous selection index for
  yield and stability.} \item{rAMGE}{The ranks of AMGE values.} \item{rY}{The
  ranks of the mean yield of genotypes.} \item{means}{The mean yield of the
  genotypes.}

  The names of the genotypes are indicated as the row names of the data
  frame.
}
\description{
\code{AMGE.AMMI} computes the Sum Across Environments of Genotype-Environment
Interaction (GEI) Modelled by AMMI (AMGE) (Sneller et al., 1997) considering
all significant interaction principal components (IPCs) in the AMMI model.
Using AMGE, the Simultaneous Selection Index for Yield and Stability (SSI) is
also calculated according to the argument \code{ssi.method}.
}
\details{
The Sum Across Environments of GEI Modelled by AMMI
(\ifelse{html}{\out{<i>AMGE</i>}}{\eqn{AMGE}}) is computed as follows:

\ifelse{html}{\out{<p style="text-align: center;"><i>AMGE =
<big>&sum;</big><sup>E</sup><sub style="line-height: 1.8; margin-left:
-1ex;">j=1</sub><big>&sum;</big><sup>N'</sup><sub style="line-height: 1.8;
margin-left: -2ex;">n=1</sub>
&lambda;<sub>n</sub>&gamma;<sub>in</sub>&delta;<sub>jn</sub></i></p>}}{\deqn{AMGE
= \sum_{j=1}^{E} \sum_{n=1}^{N'} \lambda_{n} \gamma_{in} \delta_{jn}}}

Where, \ifelse{html}{\out{<i>N'</i>}}{\eqn{N'}} is the number of significant
IPCs (number of IPC that were retained in the AMMI model via F tests);
\ifelse{html}{\out{<i>&lambda;<sub>n</sub></i>}}{\eqn{\lambda_{n}}} is the
singular value for \ifelse{html}{\out{<i>n</i>}}{\eqn{n}}th IPC and correspondingly
\ifelse{html}{\out{<i>&lambda;<sup>2</sup><sub style="line-height: 1.8;
margin-left: -1ex;">n</sub></i>}}{\eqn{\lambda_{n}^{2}}} is its eigen value;
\ifelse{html}{\out{<i>&gamma;<sub>in<sub></i>}}{\eqn{\gamma_{in}}} is the
eigenvector value for \ifelse{html}{\out{<i>i</i>}}{\eqn{i}}th genotype; and
\ifelse{html}{\out{<i>&delta;<sub>jn<sub></i>}}{\eqn{\delta{jn}}} is the
eigenvector value for the \ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th environment.
}
\examples{
library(agricolae)
data(plrv)

# AMMI model
model <- with(plrv, AMMI(Locality, Genotype, Rep, Yield, console = FALSE))

# ANOVA
model$ANOVA

# IPC F test
model$analysis

# Mean yield and IPC scores
model$biplot

# G*E matrix (deviations from mean)
array(model$genXenv, dim(model$genXenv), dimnames(model$genXenv))

# With default n (N') and default ssi.method (farshadfar)
AMGE.AMMI(model)

# With n = 4 and default ssi.method (farshadfar)
AMGE.AMMI(model, n = 4)

# With default n (N') and ssi.method = "rao"
AMGE.AMMI(model, ssi.method = "rao")

# Changing the ratio of weights for Rao's SSI
AMGE.AMMI(model, ssi.method = "rao", a = 0.43)

}
\references{
\insertRef{sneller_repeatability_1997}{AMMIStbP}
}
\seealso{
\code{\link[AMMIStbP]{SSI}}
}
