% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASI.AMMI.R
\name{ASI.AMMI}
\alias{ASI.AMMI}
\title{AMMI Stability Index}
\usage{
ASI.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"),
  a = 1)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}).}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[AMMIStbP]{SSI}}).}
}
\value{

}
\description{
\code{ASI.AMMI} computes the AMMI Stability Index (ASI) (Jambhulkar et al.,
2014; Jambhulkar et al., 2015; Jambhulkar et al., 2017) considering the first
two interaction principal components (IPCs) in the AMMI model. Using ASI, the
Simultaneous Selection Index for Yield and Stability (SSI) is also calculated
according to the argument \code{ssi.method}.
}
\details{
The AMMI Stability Index (\eqn{ASI}) is computed as follows:

\deqn{ASI = \sqrt{\left [ PC_{1}^{2} \times \theta_{1}^{2} \right ]+\left [
PC_{2}^{2} \times \theta_{2}^{2} \right ]}}

Where, \eqn{PC_{1}} and \eqn{PC_{2}} are the scores of 1st and 2nd IPCs
respectively; and \eqn{\theta_{1}} and \eqn{\theta_{2}} are percentage sum of
squares explained by the 1st and 2nd principal component interaction effect
respectively.

The Yield Stability Index (\eqn{YSI}) is computed as follows:
}
\references{
\insertRef{jambhulkar_ammi_2014}{AMMIStbP}

\insertRef{jambhulkar_genotype_2015}{AMMIStbP}

\insertRef{jambhulkar_stability_2017}{AMMIStbP}
}
\seealso{
\code{\link[AMMIStbP]{SSI}}
}
