% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTAB.AMMI.R
\name{ASTAB.AMMI}
\alias{ASTAB.AMMI}
\title{AMMI Based Stability Parameter}
\usage{
ASTAB.AMMI(model, n, alpha = 0.05)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}.}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}
}
\value{

}
\description{
\code{ASTAB.AMMI} computes the AMMI Based Stability Parameter (ASTAB) (Rao
and Prabhakaran, 2005) considering all significant interaction principal
components (IPCs) in the AMMI model. Using ASTAB, the Yield stability Index
(YSI) is also calculated.
}
\details{
The AMMI Based Stability Parameter value (\eqn{ASTAB}) is computed as
follows:

\deqn{ASTAB = \sum_{n=1}^{N'}\lambda_{n}\gamma_{in}^{2}}

Where, \eqn{N'} is the number of significant IPCAs (number of IPC that were
retained in the AMMI model via F tests); \eqn{\lambda_{n}} is the is the
singular value for IPC \eqn{n} and correspondingly \eqn{\lambda_{n}^{2}}  is
its eigen value; and \eqn{\gamma_{in}} is the eigenvector value for \eqn{i}th
genotype.

The Yield Stability Index (\eqn{YSI}) is computed as follows:

\deqn{YSI = R_{ASTAB} + R_{Y}}

Where, \eqn{R_{ASTAB}} is the ASTAB rank of the
genotype and \eqn{R_{Y}} is the mean yield rank of the genotype.
}
\references{
\insertRef{rao_use_2005}{AMMIStbP}
}
