% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVAMGE.AMMI.R
\name{AVAMGE.AMMI}
\alias{AVAMGE.AMMI}
\title{Sum Across Environments of Absolute Value of GEI Modelled by AMMI}
\usage{
AVAMGE.AMMI(model, n, alpha = 0.05)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}.}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be}
}
\value{

}
\description{
\code{AVAMGE.AMMI} computes the Sum Across Environments of Absolute Value of
GEI Modelled by AMMI (AVAMGE) (Zali et al., 2012) considering all significant
interaction principal components in the AMMI model. Using AVAMGE, the Yield
stability Index (YSI) is also calculated.
}
\details{
The Sum Across Environments of Absolute Value of GEI Modelled by AMMI
(\eqn{AVAMGE}) is computed as follows:

\deqn{AV_{(AMGE)} = \sum_{j=1}^{E} \sum_{n=1}^{N'} \left |\lambda_{n}
\gamma_{in} \delta_{jn}  \right |}

Where, \eqn{N'} is the number of significant IPCAs (number of IPC that were
retained in the AMMI model via F tests); \eqn{\lambda_{n}} is the is the
singular value for IPC \eqn{n} and correspondingly \eqn{\lambda_{n}^{2}}  is
its eigen value; \eqn{\gamma_{in}} is the eigenvector value for \eqn{i}th
genotype; and \eqn{\delta_{jn}}  is the eigenvector value for \eqn{j}th
environment.

The Yield Stability Index (\eqn{YSI}) is computed as follows:

\deqn{YSI = R_{AVAMGE} + R_{Y}}

Where, \eqn{R_{AVAMGE}} is the SIPC rank of the genotype and \eqn{R_{Y}} is
the mean yield rank of the genotype.
}
\references{
\insertRef{zali_evaluation_2012}{AMMIStbP}
}
