% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.AMMI.R
\name{DA.AMMI}
\alias{DA.AMMI}
\title{Annicchiarico's D Parameter}
\usage{
DA.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"),
  a = 1)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}).}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[AMMIStbP]{SSI}}).}
}
\value{

}
\description{
\code{DA.AMMI} computes the Annicchiarico's D Parameter values
(\eqn{\textrm{D}_{\textrm{a}}}) (Annicchiarico, 1997) considering all
significant interaction principal components (IPCs) in the AMMI model. It is
the unsquared Euclidean distance from the origin of significant IPC axes in
the AMMI model. Using \eqn{\textrm{D}_{\textrm{a}}}, the Simultaneous
Selection Index for Yield and Stability (SSI) is also calculated according to
the argument \code{ssi.method}.
}
\details{
The Annicchiarico's D Parameter value (\eqn{D_{a}}) is computed as follows:

\deqn{D_{a} = \sqrt{\sum_{n=1}^{N'}(\lambda_{n}\gamma_{in})^2}}

Where, \eqn{N'} is the number of significant IPCAs (number of IPC that were
retained in the AMMI model via F tests); \eqn{\lambda_{n}} is the singular
value for IPC \eqn{n} and correspondingly \eqn{\lambda_{n}^{2}}  is its eigen
value; and \eqn{\gamma_{in}} is the eigenvector value for \eqn{i}th genotype.

The Yield Stability Index (\eqn{YSI}) is computed as follows:
}
\references{
\insertRef{annicchiarico_joint_1997}{AMMIStbP}
}
\seealso{
\code{\link[AMMIStbP]{SSI}}
}
