% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DZ.AMMI.R
\name{DZ.AMMI}
\alias{DZ.AMMI}
\title{Zhang's D Parameter}
\usage{
DZ.AMMI(model, n, alpha = 0.05)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}.}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}
}
\value{

}
\description{
\code{DZ.AMMI} computes the Zhang's D Parameter values or AMMI statistic
coefficient or AMMI distance or AMMI stability index
(\eqn{\textrm{D}_{\textrm{z}}}) (Zhang, 1998) considering all significant
interaction principal components (IPCs) in the AMMI model. It is the distance
of IPC  point  from  origin  in space. Using \eqn{\textrm{D}_{\textrm{z}}},
the Yield stability Index (YSI) is also calculated.
}
\details{
The Zhang's D Parameter value (\eqn{D_{z}}) is computed as follows:

\deqn{D_{z} = \sqrt{\sum_{n=1}^{N'}\gamma_{in}^{2}}}

Where, \eqn{N'} is the number of significant IPCAs (number of IPC that were
retained in the AMMI model via F tests); and \eqn{\gamma_{in}} is the
eigenvector value for \eqn{i}th genotype.

The Yield Stability Index (\eqn{YSI}) is computed as follows:

\deqn{YSI = R_{D_{z}} + R_{Y}}

Where, \eqn{R_{D_{z}}} is the \eqn{\textrm{D}_{\textrm{z}}} rank of the
genotype and \eqn{R_{Y}} is the mean yield rank of the genotype.
}
\references{
\insertRef{zhang_analysis_1998}{AMMIStbP}
}
