% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DZ.AMMI.R
\name{DZ.AMMI}
\alias{DZ.AMMI}
\title{Zhang's D Parameter}
\usage{
DZ.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"),
  a = 1)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}).}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[AMMIStbP]{SSI}}).}
}
\value{

}
\description{
\code{DZ.AMMI} computes the Zhang's D Parameter values or AMMI statistic
coefficient or AMMI distance or AMMI stability index
(\eqn{\textrm{D}_{\textrm{z}}}) (Zhang, 1998) considering all significant
interaction principal components (IPCs) in the AMMI model. It is the distance
of IPC  point  from  origin  in space. Using \eqn{\textrm{D}_{\textrm{z}}},
the Simultaneous Selection Index for Yield and Stability (SSI) is also
calculated according to the argument \code{ssi.method}.
}
\details{
The Zhang's D Parameter value (\eqn{D_{z}}) is computed as follows:

\deqn{D_{z} = \sqrt{\sum_{n=1}^{N'}\gamma_{in}^{2}}}

Where, \eqn{N'} is the number of significant IPCAs (number of IPC that were
retained in the AMMI model via F tests); and \eqn{\gamma_{in}} is the
eigenvector value for \eqn{i}th genotype.
}
\references{
\insertRef{zhang_analysis_1998}{AMMIStbP}
}
\seealso{
\code{\link[AMMIStbP]{SSI}}
}
