% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV.AMMI.R
\name{EV.AMMI}
\alias{EV.AMMI}
\title{Averages of the Squared Eigenvector Values}
\usage{
EV.AMMI(model, n, alpha = 0.05)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}.}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}
}
\value{

}
\description{
\code{EV.AMMI} computes the Sums of the Averages of the Squared Eigenvector
Values (EV) (Zobel, 1994) considering all significant interaction principal
components (IPCs) in the AMMI model. Using EV, the Yield stability Index (YSI) is
also calculated.
}
\details{
The Averages of the Squared Eigenvector Values (\eqn{EV}) is computed as
follows:

\deqn{EV = \sum_{n=1}^{N'}\frac{\gamma_{in}^2}{N'}}

Where, \eqn{N'} is the number of significant IPCs (number of IPC that were
retained in the AMMI model via F tests); and \eqn{\gamma_{in}} is the
eigenvector value for \eqn{i}th genotype.

The Yield Stability Index (\eqn{YSI}) is computed as follows:

\deqn{YSI = R_{EV} + R_{Y}}

Where, \eqn{R_{EV}} is the EV rank of the genotype and \eqn{R_{Y}} is the
mean yield rank of the genotype.
}
\references{
\insertRef{zobel_stress_1994}{AMMIStbP}
}
