% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV.AMMI.R
\name{EV.AMMI}
\alias{EV.AMMI}
\title{Averages of the Squared Eigenvector Values}
\usage{
EV.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"),
  a = 1)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}).}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[AMMIStbP]{SSI}}).}
}
\value{

}
\description{
\code{EV.AMMI} computes the Sums of the Averages of the Squared Eigenvector
Values (EV) (Zobel, 1994) considering all significant interaction principal
components (IPCs) in the AMMI model. Using EV, the Simultaneous Selection
Index for Yield and Stability (SSI) is also calculated according to the
argument \code{ssi.method}.
}
\details{
The Averages of the Squared Eigenvector Values (\eqn{EV}) is computed as
follows:

\deqn{EV = \sum_{n=1}^{N'}\frac{\gamma_{in}^2}{N'}}

Where, \eqn{N'} is the number of significant IPCs (number of IPC that were
retained in the AMMI model via F tests); and \eqn{\gamma_{in}} is the
eigenvector value for \eqn{i}th genotype.
}
\references{
\insertRef{zobel_stress_1994}{AMMIStbP}
}
\seealso{
\code{\link[AMMIStbP]{SSI}}
}
