% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASV.AMMI.R
\name{MASV.AMMI}
\alias{MASV.AMMI}
\title{Modified AMMI Stability Value}
\usage{
MASV.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"),
  a = 1)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}).}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[AMMIStbP]{SSI}}).}
}
\value{
A data frame with the following columns:  \item{MASV}{The MASV
  values.} \item{SSI}{The computed values of simultaneous selection index for
  yield and stability.} \item{rMASV}{The ranks of MASV values.} \item{rY}{The
  ranks of the mean yield of genotypes.} \item{means}{The mean yield of the
  genotypes.}

  The names of the genotypes are indicated as the row names of the data
  frame.
}
\description{
\code{MASV.AMMI} computes the Modified AMMI Stability Value (MASV) (Zali et
al., 2012; Please see \strong{Note}) from a modified formula of AMMI
Stability Value (ASV) (Purchase et al. 1997). This formula calculates AMMI
stability value considering all significant interaction principal components
(IPCs) in the AMMI model. Using MASV, the Simultaneous Selection Index for
Yield and Stability (SSI) is also calculated according to the argument
\code{ssi.method}.
}
\details{
The Modified AMMI Stability Value
(\ifelse{html}{\out{<i>MASV</i>}}{\eqn{MASV}}) is computed as follows:

\ifelse{html}{\out{<p style="text-align: center;"><i>MASV = &radic;<span
style="text-decoration:overline">&sum;<sup>N'-1</sup><sub style="line-height:
1.8; margin-left: -4ex;">n=1</sub>((SSIPC<sub>n</sub> &frasl;
SSIPC<sub>n+1</sub>) &times; PC<sub>1</sub>)<sup>2</sup> +
(PC<sub>N'</sub>)<sup>2</sup></span></i></p>}}{\deqn{MASV =
\sqrt{\sum_{n=1}^{N'-1}\left (\frac{SSIPC_{n}}{SSIPC_{n+1}} \times PC_{n}
\right )^2   + \left (PC_{N'}  \right )^2}}}

Where, \ifelse{html}{\out{<i>SSIPC<sub>1</sub></i>}}{\eqn{SSIPC_{1}}},
\ifelse{html}{\out{<i>SSIPC<sub>2</sub></i>}}{\eqn{SSIPC_{2}}},
\ifelse{html}{\out{&hellip;}}{\eqn{\cdots}},
\ifelse{html}{\out{<i>SSIPC<sub>n</sub></i>}}{\eqn{SSIPC_{n}}} are the sum of
squares of the 1st, 2nd, ..., and \ifelse{html}{\out{<i>n</i>}}{\eqn{n}}th
IPC; and \ifelse{html}{\out{<i>PC<sub>1</sub></i>}}{\eqn{PC_{1}}},
\ifelse{html}{\out{<i>PC<sub>2</sub></i>}}{\eqn{PC_{2}}},
\ifelse{html}{\out{&hellip;}}{\eqn{\cdots}},
\ifelse{html}{\out{<i>PC<sub>n</sub></i>}}{\eqn{PC_{n}}} are the scores of
1st, 2nd, ..., and \ifelse{html}{\out{<i>n</i>}}{\eqn{n}}th IPC.
}
\note{
In Zali et al., (2012), the formula for both AMMI stability value (ASV)
  was found to be erroneus, when compared with the original publications
  (Purchase 1997; Purchase et al., 1999; Purchase et al., 2000).

  \strong{ASV (Zali et al., 2012)} \ifelse{html}{\out{<p style="text-align:
  center;"><i>ASV = &radic;<span
  style="text-decoration:overline">(SSIPC<sub>1</sub> &frasl;
  SSIPC<sub>2</sub>) &times; (PC<sub>1</sub>)<sup>2</sup> +
  (PC<sub>2</sub>)<sup>2</sup></span></i></p>}}{\deqn{ASV = \sqrt{\left (
  \frac{SSIPC_{1}}{SSIPC_{2}} \right ) \times (PC_{1})^2   + \left (PC_{2}
  \right )^2}}}

  \strong{ASV (Purchase 1997; Purchase et al., 1999; Purchase et al., 2000)}
  \ifelse{html}{\out{<p style="text-align: center;"><i>ASV = &radic;<span
  style="text-decoration:overline">((SSIPC<sub>1</sub> &frasl;
  SSIPC<sub>2</sub>) &times; PC<sub>1</sub>)<sup>2</sup> +
  (PC<sub>2</sub>)<sup>2</sup></span></i></p>}}{\deqn{ASV = \sqrt{\left
  (\frac{SSIPC_{1}}{SSIPC_{2}} \times PC_{1}  \right )^2   + \left (PC_{2}
  \right )^2}}}

  The authors believe that the proposed Modified AMMI stability value (MASV)
  in Zali et al., (2012) is also erroneous and have implemented the corrected
  one in \code{MASV.AMMI}.

  \strong{MASV (Zali et al., 2012)} \ifelse{html}{\out{<p style="text-align:
  center;"><i>MASV = &radic;<span
  style="text-decoration:overline">&sum;<sup>N'-1</sup><sub
  style="line-height: 1.8; margin-left: -4ex;">n=1</sub>(SSIPC<sub>n</sub>
  &frasl; SSIPC<sub>n+1</sub>) &times; (PC<sub>1</sub>)<sup>2</sup> +
  (PC<sub>N'</sub>)<sup>2</sup></span></i></p>}}{\deqn{MASV =
  \sqrt{\sum_{n=1}^{N'-1}\left ( \frac{SSIPC_{n}}{SSIPC_{n+1}} \right )
  \times (PC_{n})^2   + \left (PC_{N'}  \right )^2}}}
}
\examples{
library(agricolae)
data(plrv)

# AMMI model
model <- with(plrv, AMMI(Locality, Genotype, Rep, Yield, console = FALSE))

# ANOVA
model$ANOVA

# IPC F test
model$analysis

# Mean yield and IPC scores
model$biplot

# G*E matrix (deviations from mean)
array(model$genXenv, dim(model$genXenv), dimnames(model$genXenv))

# With default n (N') and default ssi.method (farshadfar)
MASV.AMMI(model)

# With n = 4 and default ssi.method (farshadfar)
MASV.AMMI(model, n = 4)

# With default n (N') and ssi.method = "rao"
MASV.AMMI(model, ssi.method = "rao")

# Changing the ratio of weights for Rao's SSI
MASV.AMMI(model, ssi.method = "rao", a = 0.43)

}
\references{
\insertRef{purchase_parametric_1997}{AMMIStbP}

\insertRef{purchase_use_1999}{AMMIStbP}

\insertRef{purchase_genotype_2000}{AMMIStbP}

\insertRef{zali_evaluation_2012}{AMMIStbP}
}
\seealso{
\code{\link[AMMIStbP]{SSI}}
}
