% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASV.AMMI.R
\name{MASV.AMMI}
\alias{MASV.AMMI}
\title{Modified AMMI Stability Value}
\usage{
MASV.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"),
  a = 1)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}).}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[AMMIStbP]{SSI}}).}
}
\value{

}
\description{
\code{MASV.AMMI} computes the Modified AMMI Stability Value (MASV) (Zali et
al., 2012; Please see \strong{Note}) from a modified formula of AMMI
Stability Value (ASV) (Purchase et al. 1997). This formula calculates AMMI
stability value considering all significant interaction principal components
(IPCs) in the AMMI model. Using MASV, the Simultaneous Selection Index for
Yield and Stability (SSI) is also calculated according to the argument
\code{ssi.method}.
}
\details{
The Modified AMMI Stability Value (\eqn{MASV}) is computed as follows:

\deqn{MASV = \sqrt{\sum_{n=1}^{N'-1}\left (\frac{SSIPC_{n}}{SSIPC_{n+1}}
\times PC_{n}  \right )^2   + \left (PC_{N'}  \right )^2}}

Where, \eqn{SSIPC_{1}}, \eqn{SSIPC_{2}}, \eqn{\cdots}, \eqn{SSIPC_{n}} are
the sum of squares of the 1st, 2nd, ..., and \eqn{n}th IPC; and \eqn{PC_{1}},
\eqn{PC_{2}}, \eqn{\cdots}, \eqn{PC_{n}} are the scores of 1st, 2nd, ..., and
\eqn{n}th IPC.

The Yield Stability Index (\eqn{YSI}) is computed as follows:

\deqn{YSI = R_{MASV} + R_{Y}}

Where, \eqn{R_{MASV}} is the MASV rank of the genotype and \eqn{R_{Y}} is the
mean yield rank of the genotype.
}
\note{
In Zali et al., (2012), the formula for both AMMI stability value (ASV)
  was found to be erroneus, when compared with the original publications
  (Purchase 1997; Purchase et al., 1999; Purchase et al., 2000).

  \strong{ASV (Zali et al., 2012)} \deqn{ASV = \sqrt{\left (
  \frac{SSIPC_{1}}{SSIPC_{2}} \right ) \times (PC_{1})^2   + \left (PC_{2}
  \right )^2}}

  \strong{ASV (Purchase 1997; Purchase et al., 1999; Purchase et al., 2000)}
  \deqn{ASV = \sqrt{\left (\frac{SSIPC_{1}}{SSIPC_{2}} \times PC_{1}  \right
  )^2   + \left (PC_{2}  \right )^2}}

  The authors believe that the proposed Modified AMMI stability value (MASV)
  in Zali et al., (2012) is also erroneous and have implemented the corrected
  one in \code{MASV.AMMI}.

  \strong{MASV (Zali et al., 2012)} \deqn{MASV = \sqrt{\sum_{n=1}^{N'-1}\left
  ( \frac{SSIPC_{n}}{SSIPC_{n+1}} \right ) \times (PC_{n})^2   + \left
  (PC_{N'}  \right )^2}}
}
\references{
\insertRef{purchase_parametric_1997}{AMMIStbP}

\insertRef{purchase_use_1999}{AMMIStbP}

\insertRef{purchase_genotype_2000}{AMMIStbP}

\insertRef{zali_evaluation_2012}{AMMIStbP}
}
\seealso{
\code{\link[AMMIStbP]{SSI}}
}
