% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPC.AMMI.R
\name{SIPC.AMMI}
\alias{SIPC.AMMI}
\title{Sums of the Absolute Value of the IPC Scores}
\usage{
SIPC.AMMI(model, n, alpha = 0.05)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}.}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be}
}
\value{

}
\description{
\code{SIPC.AMMI} computes the Sums of the Absolute Value of the IPC Scores
(ASI) (Sneller et al., 1997) all significant interaction principal components
in the AMMI model. Using SIPC, the Yield stability Index (YSI) is also
calculated.
}
\details{
The Sums of the Absolute Value of the IPC Scores (\eqn{SIPC}) is computed as
follows:

\deqn{SIPC = \sum_{n=1}^{N'} \left | \lambda_{n}^{0.5}\gamma_{in} \right |}

OR

\deqn{SIPC = \sum_{n=1}^{N'}\left | PC_{n} \right |}

Where, \eqn{N'} is the number of significant IPCAs (number of IPC that were
retained in the AMMI model via F tests); \eqn{\lambda_{n}} is the is the singular
value for IPC \eqn{n} and correspondingly \eqn{\lambda_{n}^{2}}  is its eigen value;
\eqn{\gamma_{in}} is the eigenvector value for \eqn{i}th genotype; and \eqn{PC_{1}},
\eqn{PC_{2}}, \eqn{\cdots}, \eqn{PC_{n}} are the scores of 1th, 2th, ..., and \eqn{n}th IPC.

The closer the SIPC scores are to zero, the more stable the genotypes are
across test environments.

The Yield Stability Index (\eqn{YSI}) is computed as follows:

\deqn{YSI = R_{SIPC} + R_{Y}}

Where, \eqn{R_{SIPC}} is the SIPC rank of the genotype and \eqn{R_{Y}} is
the mean yield rank of the genotype.
}
\references{
\insertRef{sneller_repeatability_1997}{AMMIStbP}
}
