% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPC.AMMI.R
\name{SIPC.AMMI}
\alias{SIPC.AMMI}
\title{Sums of the Absolute Value of the IPC Scores}
\usage{
SIPC.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"),
  a = 1)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}.}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}

\item{ssi.method}{The method for the computation of simultaneous selection
index. Either \code{"farshadfar"} or \code{"rao"} (See
\code{\link[AMMIStbP]{SSI}}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See
\code{\link[AMMIStbP]{SSI}}).}
}
\value{

}
\description{
\code{SIPC.AMMI} computes the Sums of the Absolute Value of the IPC Scores
(ASI) (Sneller et al., 1997) considering all significant interaction
principal components (IPCs) in the AMMI model. Using SIPC, the Simultaneous
Selection Index for Yield and Stability (SSI) is also calculated according to
the argument \code{ssi.method}.
}
\details{
The Sums of the Absolute Value of the IPC Scores (\eqn{SIPC}) is computed as
follows:

\deqn{SIPC = \sum_{n=1}^{N'} \left | \lambda_{n}^{0.5}\gamma_{in} \right |}

OR

\deqn{SIPC = \sum_{n=1}^{N'}\left | PC_{n} \right |}

Where, \eqn{N'} is the number of significant IPCs (number of IPC that were
retained in the AMMI model via F tests); \eqn{\lambda_{n}} is the is the
singular value for IPC \eqn{n} and correspondingly \eqn{\lambda_{n}^{2}}  is
its eigen value; \eqn{\gamma_{in}} is the eigenvector value for \eqn{i}th
genotype; and \eqn{PC_{1}}, \eqn{PC_{2}}, \eqn{\cdots}, \eqn{PC_{n}} are the
scores of 1th, 2th, ..., and \eqn{n}th IPC.

The closer the SIPC scores are to zero, the more stable the genotypes are
across test environments.
}
\references{
\insertRef{sneller_repeatability_1997}{AMMIStbP}
}
\seealso{
\code{\link[AMMIStbP]{SSI}}
}
