% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSI.R
\name{SSI}
\alias{SSI}
\title{Simultaneous Selection Indices for Yield and Stability}
\usage{
SSI(y, sp, gen, method = c("farshadfar", "rao"), a = 1)
}
\arguments{
\item{y}{A numeric vector of the mean yield/performance of genotypes.}

\item{sp}{A numeric vector of the stability paramter/index of the genotypes.}

\item{gen}{A character vector of the names of the genotypes.}

\item{method}{The method for the computation of simultaneous selection index.
Either \code{"farshadfar"} or \code{"rao"} (See \strong{Details}).}

\item{a}{The ratio of the weights given to the stability components for
computation of SSI when \code{method = "rao"} (See \strong{Details}).}
}
\description{
\code{SSI} computes the Simultaneous Selection Index for Yield and Stability
(SSI) according to the methods specified in the argument \code{method}.
}
\details{
The SSI proposed by Rao and Prabhakaran (2005) (\eqn{I_{i}}) is computed as
follows:

\deqn{I_{i} = \frac{\overline{Y}_{i}}{\overline{Y}_{..}} + \alpha
\frac{\frac{1}{SP_{i}}}{\frac{1}{T}\sum_{i=1}^{T}\frac{1}{SP_{i}}}}

Where \eqn{SP_{i}} is the stability measure of the \eqn{i}th genotype under
AMMI procedure;  \eqn{Y_{i}} is mean performance of \eqn{i}th genotype;
\eqn{Y_{..}} is the overall mean; \eqn{T} is the number of genotypes under
test and \eqn{\alpha} is the ratio of the weights given to the stability
components (\eqn{w_{2}}) and yield (\eqn{w_{1}}) with a restriction that
\eqn{w_{1} + w_{2} = 1}. The weights can be specified as required.

\tabular{rrr}{
\strong{\eqn{\alpha}} \tab \strong{\eqn{w_{1}}} \tab \strong{\eqn{w_{2}}}\cr
1.00 \tab 0.5 \tab 0.5\cr
0.67 \tab 0.6 \tab 0.4\cr
0.43 \tab 0.7 \tab 0.3\cr
0.25 \tab 0.8 \tab 0.2\cr
}

The SSI proposed by Farshadfar (2008) is called the Genotype stability index
(GSI) or Yield stability index (YSI) (Farshadfar et al., 2011) and is
computed by summation of the ranks of the stability index/parameter and the
ranks of the mean yields.

\deqn{GSI = YSI = R_{SP} + R_{Y}}

Where, \eqn{R_{SP}} is the stability parameter/index rank of the genotype and
\eqn{R_{Y}} is the mean yield rank of the genotype.
}
\references{
\insertRef{rao_use_2005}{AMMIStbP}

\insertRef{farshadfar_incorporation_2008}{AMMIStbP}

\insertRef{farshadfar_ammi_2011}{AMMIStbP}
}
