% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZA.AMMI.R
\name{ZA.AMMI}
\alias{ZA.AMMI}
\title{Absolute value of the Relative Contribution of IPCs to the Interaction}
\usage{
ZA.AMMI(model, n, alpha = 0.05)
}
\arguments{
\item{model}{The AMMI model (An object of class \code{AMMI} generated by
\code{\link[agricolae]{AMMI}}.}

\item{n}{The number of principal components to be considered for computation.
The default value is the number of significant IPCs.}

\item{alpha}{Type I error probability (Significance level) to be considered
to identify the number of significant IPCs.}
}
\value{

}
\description{
\code{ZA.AMMI} computes the Absolute value of the Relative Contribution of
IPCs to the Interaction (Za) (Zali, 2012) considering all significant
interaction principal components (IPCs) in the AMMI model. Using Za, the
Yield stability Index (YSI) is also calculated.
}
\details{
The Absolute value of the Relative Contribution of IPCs to the Interaction
(\eqn{D_{Za}}) is computed as follows:

\deqn{Za = \sum_{i=1}^{N'}\left | \theta_{n}\gamma_{in} \right |}

Where, \eqn{N'} is the number of significant IPCAs (number of IPC that were
retained in the AMMI model via F tests); and \eqn{\gamma_{in}} is the
eigenvector value for \eqn{i}th genotype; and \eqn{\theta_{n}} is the
percentage sum of squares explained by \eqn{n}th principal component
interaction effect.

The Yield Stability Index (\eqn{YSI}) is computed as follows:

\deqn{YSI = R_{Za} + R_{Y}}

Where, \eqn{R_{Za}} is the Za rank of the
genotype and \eqn{R_{Y}} is the mean yield rank of the genotype.
}
\references{
\insertRef{zali_evaluation_2012}{AMMIStbP}
}
